requirements_311.txt
# Core frameworks
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.24.0
python-dotenv==1.0.0
pydantic==1.10.8
httpx==0.24.1

# Agent frameworks
agno==0.1.1
langchain==0.0.200
crewai==0.11.2

# Vector database
faiss-cpu==1.7.4
pinecone-client==2.2.1

# Data processing
pandas==2.0.2
numpy==1.24.3
yfinance==0.2.18
alpha_vantage==2.3.1
beautifulsoup4==4.12.2
requests==2.31.0

# NLP & ML
transformers==4.30.2
sentence-transformers==2.2.2
openai==0.27.8

# Voice processing
pyttsx3==2.90
SpeechRecognition==3.10.0

# Testing
pytest==7.3.1
